Progress towards option to pick `orographic_model`. See extended message
This commit represents an intermedia step towards the ability to pick
the orographic model. Here we allow an extra option through the
variable `slopeAndAspectMode`, which can be either "compute" or
"download" and it controls how the slope and aspect maps are obtained.

This was done as a simple study to determine the best way to obtain
these maps. This commit includes a notebook containing the study. This
commit includes several lines of code that will be removed in the
following commits, intended only for the study (e.g. the variable
mentioned above).