Clean up code with sim_seed -- see note
Currently, the initializer for the multiprocessing.Pool is *almost*
working correctly--I can pull out a worker ID, but unfortunately it is
not 1..num_cores. Instead, the id is incremented with every new worker
instance, i.e., with each subsequent call to mp.Pool.map(). Moreover,
the order in which workers are called is not fixed.